fix data gr not found

+++ src/app/modules/DeliveryMonitoring/pages/Termin/DeliveryOrder.js/components/ModalSubmit.js
@@ -109,7 +109,7 @@ const ModalSubmit = ({
           </TableHead>
           <TableBody>
             {tempOrderItems?.map((item, index) => {
-              let qtyAvail = parseInt(item?.qty_delivery_available);
+              let qtyAvail = parseFloat(item?.qty_delivery_available);
               return (
                 <TableRow key={index}>
                   <TableCell>
